+++ R/check_data.R
@@ -97,7 +97,7 @@ basic_info <- function(df, y = NULL, time = NULL, status = NULL, verbose = TRUE)
       '\n\nWith the target described by ', target, '\n\n', sep = '', verbose = verbose)
   str <- capture.output(cat('**The dataset has ', nrow(df), ' observations and ', ncol(df),
                             ' columns which names are: **\n\n', paste0(colnames(df), sep='; '),
-                            '\n\n**With the target described by ', target, '\n \n', sep = ''))
+                            '\n\n **With the target described by** ', target, '\n \n', sep = ''))
   return(str)
 }
 
@@ -136,7 +136,7 @@ check_static <- function(df, verbose = TRUE) {
   } else {
     verbose_cat(crayon::red('\u2716'), ' Static columns are: \n ', dominator_cols, '\n\n', sep = '', verbose = verbose)
     verbose_cat(crayon::red('\u2716'), ' With dominating values: \n', ' ',  dominator_vals, '\n ', sep = '', verbose = verbose)
-    str <- capture.output(cat('** Static columns are: **', dominator_cols, '\n\n', sep = ''))
+    str <- capture.output(cat('**Static columns are: **', dominator_cols, '\n\n', sep = ''))
     str <- c(str, capture.output(cat('**With dominating values: **', dominator_vals,
                                      '\n\n', sep = '')))
   }
@@ -379,7 +379,7 @@ check_cor <- function(df, y = NULL, time = NULL, status = NULL, verbose = TRUE) 
         if (i != j && strong_V_cor) {
           if (no_cor_fct) {
             verbose_cat(crayon::red('\u2716'), 'Strongly correlated, by Crammer\'s V rank, pairs of categorical values are: \n', verbose = verbose)
-            str <- c(str, capture.output(cat('\n', '** Strongly correlated, by Crammer\'s V rank, pairs of categorical values are: **\n\n')))
+            str <- c(str, capture.output(cat('\n', '**Strongly correlated, by Crammer\'s V rank, pairs of categorical values are: **\n\n')))
             no_cor_fct = FALSE
           }
           verbose_cat(' ', fct_names[i], ' - ', fct_names[j], ': ', cor_fct[i, j], ';\n', sep = '', verbose = verbose)